+++ packages/pak-files-parser/src/app/generators/npcs/npc-db.generator.ts
@@ -100,9 +100,9 @@ export class NPCDbGenerator extends BaseGenerator<RawNPC, NPC> {
         return defaultAppearance;
     }
 
-    private static getNPCAppearances(dbItem: RawNPC, itemKey: string) {
+    private static getNPCAppearances(dbItem: RawNPC, itemKey: string, newAppearances = false) {
         let npcAppearances: Record<string, RawNpcAppearances> = {}
-        const {index, fileName} = extractOutfitPortraitsLocation(dbItem, itemKey);
+        const {index, fileName} = extractOutfitPortraitsLocation(dbItem, itemKey, newAppearances);
 
         try {
             npcAppearances = readAsset<Datatable<RawNpcAppearances>[]>(fileName)[+index].Rows;
@@ -119,23 +119,14 @@ export class NPCDbGenerator extends BaseGenerator<RawNPC, NPC> {
         const defaultAppearance = NPCDbGenerator.formatRawAppearances(itemKey, npcAppearances);
         appearances.push({appearances: defaultAppearance});
 
-        NPCDbGenerator.AdditionalNpcAppearances
-            .filter(mapping => mapping.npcKey === itemKey)
-            .forEach(mapping => {
-
-                // set to null so it doenst auto-detect, maybe FIXME ?
-                const additionalAppearance = NPCDbGenerator.getNPCAppearances({
-                    ...dbItem,
-                    portraitsDT: null
-                }, mapping.outfitKey);
-                const formattedAdditionalAppearance = NPCDbGenerator.formatRawAppearances(itemKey, additionalAppearance);
-
-                appearances.push({
-                    appearanceCategory: mapping.appearanceName,
-                    appearances: formattedAdditionalAppearance
-                });
-            })
-
+        if (dbItem.newFormAppearance.appearanceDT) {
+            const newNpcAppearances = NPCDbGenerator.getNPCAppearances(dbItem, itemKey, true);
+            const newAppearance = NPCDbGenerator.formatRawAppearances(itemKey, newNpcAppearances);
+            appearances.push({
+                appearanceCategory: NPCDbGenerator.AdditionalNpcAppearances.find(acctionalNpcAppearance => acctionalNpcAppearance.npcKey === itemKey)?.appearanceName ?? 'New form appearances',
+                appearances: newAppearance
+            });
+        }
 
         const foundImages = [...NPCDbGenerator.filePaths.appearances].map(s => s.image).map(s => fg.convertPathToPattern(s));
         const fullPath = config.source.portraitsPath + `/**/${fg.escapePath(itemKey)}/**/Po*traits/**/*.png`;
@@ -159,7 +150,7 @@ export class NPCDbGenerator extends BaseGenerator<RawNPC, NPC> {
 
         let headerPortraitFileName: string | null = null
         let customHead;
-        const portraitPath = unifyInternalPath(dbItem.Portrait.AssetPathName).split('.')[0];
+        const portraitPath = unifyInternalPath(dbItem.defaultAppearance.halfBodyPortrait.AssetPathName).split('.')[0];
 
         const sourceImagePath = path.join(environment.assetPath, portraitPath + '.png');
         const guessedPath = path.join(environment.assetPath, 'ProjectCoral', 'Content', 'ProjectCoral', 'Textures', 'UI', 'NPCHeadPortraits', 'T_Relationship' + itemKey + '.png');