+++ packages/guide/src/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ### TBD
 
+### Changed
+
+- Updated beta data to v1.3-366 (Steam version)
+- Updated live data to v1.2-1246 (Steam version)
+- Sort museum checklist by journal order
+- Improved loading of NPC when called directly via url to reduce "No NPC found" errors
+- Improved detection of character images to show more portraits
+
 #### Added
 
 - Added options to import/export user data